pyproject.toml
[tool.poetry]
name = "obsidian-concierge"
version = "0.1.0"
description = "Obsidianユーザーのための知識管理アシスタント"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
gradio = "^4.19"
chromadb = "^0.4.22"
agno = "^0.1.0"
pyyaml = "^6.0.1"
langchain = "^0.1.4"
llama-index = "^0.9.11"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true